Change return types of headerPrevHash and blockPrevHash
Before this change, getting the previous hash for epoch 0, slot 0
returned the hash of the genesis configuration coerced into a `HeaderHash`.
This was somewhat unprincipled.

Now these functions return a 'Maybe HeaderHash' where the returned
value is a `Just` for slots other than epoch 0, slot 0 and `Nothing`
when there is no previous `HeaderHash`.

The old behaviour of these two functions are kept as `headerPrevHashGenesis`
and `blockPrevHashGenesis` because other code depends on this behaviour.